Puts and prints and gets moved to the main.rb
